Adding dynamically generated google map to park page. Adding marker based on current park page

+++ scripts/js/parkMap.js
@@ -68,8 +68,8 @@ function populatePark(park)
 {
     // Fill in the Park struct
     Park = {
-        lat: parseInt(park['latitude']),
-        lng: parseInt(park['longitude'])
+        lat: parseFloat(park['latitude']),
+        lng: parseFloat(park['longitude'])
     };
 }
